Revert changes of test message bus
In https://github.com/cookpad/streamy/pull/74 we refactored
test message bus to handle delivering of messages based on the
priority. This changes made the test message bus flaky
so we're reverting it and also it did not make much sense to
implement functionality in a test helper for this use case.